Initial JSON version
Feature branch created that moves all inputs out of the packages and
into JSON files, and creates output JSON files that include the actuals.
A new table is added to store the input and output JSON files by package
and procedure. The output files can be used as inputs to a Nodejs
program, recently added to GitHub, to produce result reports formatted
in both HTML and text. The input JSON files are read into the new table
at installation time, and read from the table thereafter. The Nodejs
project includes the formatted reports for this Oracle project. The
output JSON files are written to Oracle directory input_dir (and the
input JSON files are read from there), but I have copied them into the
project oracle root for reference.